Add a "sensor lock" to mutex sensor access (WIP)
This is in response to aaaaalbert/sensibility-testbed#19, "Stopping
one sensor-reading sandbox crashes all native parts".

Using the sensor lock, we ensure that `harshexit` does not send
termination signals to the sandbox while a sensor call is in progress.
* The new `repysensors` module implements a generic wrapper that uses
  the sensor lock;
* We wrap the `sensors.get_acceleration` call in `namespace.py`;
* `harshexit` blocks until any pending sensor calls have returned,
  and only then issues an `os.kill`.

Now, two vessels that both access the accelerometer can be started
and stopped independent from each other and the rest of the app.

If the fix turns out to work sufficiently fast, I'll go and wrap
the other sensor calls likewise.