Avoid circular import problem with `emulmisc`
I do this by removing `emulmisc` from `repysensor`'s dependencies,
and thus dropping the latter's use of `emulmisc.createlock` (for
`threading.Lock()`) and `emulmisc.getruntime` (dropped without
replacement).

NOTE: This new circular import problem surfaced in the seattleinstaller.

`repysensors` used `emulmisc.getruntime` previously to sync sensor
event timestamps with the start time of the vessel. Thus, when the
sandbox received a sensor value, it would know quite precisely when
in the timeline of the vessel that sensor event had occurred.

Without `emulmisc.getruntime`, the sensor event time is just rebased
to whenever `repysensors` is imported, which is still within tens of
milliseconds of the vessel's `getruntime` function.