WIP: Get Repy sensor interfaces from `repysensors`
(This causes changes to `namespace`.)

`repysensors` in turn imports all of the lower-level sensor/output/...
implementations, and puts locks around accesses. The locks exist so
that `harshexit` is sure to not kill the sandbox while JNI sensor
calls are in progress.

Danger, if a sensor returns `None`, then the `refine_*` wrappers
will break currently!