Add `pick2_wrap` function to `repysensors`
This is yet another wrapper function, this time to return a function
that returns element #2 (which is the third element) from the list
of values returned by a sensor call.

Use it to drop timestamps and uninterpretable trailing values from
sensor readings.

As a consequence, remove the "custom" wrappers for `get_light`,
`get_proximity` and similar that essentially had the same purpose.