Fix aaaaalbert/sensibility-testbed#15
"Sanitize `portable_popen`'s callargs"

This commit ensures that all of `portable_popen`'s callargs are strings,
and that "not too many" (an arbitrary 100 of 512 possible) arguments are
supplied.

aaaaalbert/sensibility-testbed#14 mentioned that a many-args `seash`
command could be used to crash the app. With the fix applied in this
commit, only the nodemanager log will contain a trace of a violation:
```
OSError: Invalid argument. portable_popen's argument list is restricted to 100 entries.
```

Also, I fixed a blanket `except` where excepting an `ImportError` only
makes more sense (and is required now, because the updated code may
and should raise exceptions to its caller).