Fix mistaken "Monitor death" notification
When `nonportable.py/resource_monitor` can't access the monitored
process's procfs entry anymore, an `IOError` is raised which used
to cause the grim notification mentioned above.

In fact, the procfs entry disappears when the monitored process
ends by itself or is stopped by the nodemanager, which is fine.

Therefore, adapt the monitoring / exception logic to expect
to see this happen.