Override `linux_api._process_stat_file` on Android
(This is needed due to aaaaalbert/sensibility-testbed#30, the
monitoring process can't access the monitored process's procfs
entry.)

When we build our app's APK in release mode (which disables the
`debuggable` property), the sandbox process and its resource
consumption monitor become separated so strictly that the monitor
cannot access the sandbox's `procfs` entry anymore, and thus
cannot tally its CPU consumption.

If built `debuggable`, this problem does not arise. I have yet to
see a documented mention of this side effect of a non-debug build.

This commit adds an override of `linux_api._process_stat_file`
to `android_api.py`. It checks whether we are looking at the
monitor's `procfs` entry (which the monitor can access), or the
monitored process's. Fortunately, both the monitoring and the
monitored process `fork` off of the same process, so we can
pass a file handle to the monitored process's `procfs` entry
that is still valid after the monitor forks off.

Forking happens in `nonportable.py`, so this is where the file
handle is configured.

Also in `nonportable`, add information on the underlying exception
to the infamous "Monitor process died" message.